requirements.txt
alembic==1.3.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
praw==7.0.0
prawcore==1.3.0
pycparser==2.19
pyodbc==4.0.27
pypyodbc==1.3.4.3
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.2.0
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
tornado==6.0.3
update-checker==0.17
urllib3==1.25.7
websocket-client==0.57.0
Werkzeug==0.16.0
